GCP/CloudFunctions/requirements.txt
functions-framework==3.*
pandas
numpy
google-cloud-storage
google-cloud-bigquery
google-cloud
fsspec
gcsfs
pyarrow